Update AuthContext:
We'll need a new function in AuthContext like updateMentorAvailability(newSlots: AvailabilitySlot[]) which will update the specific mentor's data in MOCK_USERS and update the local user state if the logged-in user is the mentor being updated. This will also increment the bookingsVersion to ensure other components (like the schedule page) refresh if needed.
This would make the mentor side of the scheduling feature much more functional and interactive.